+++ R/internal_dag.R
@@ -1,6 +1,7 @@
 #' Call the `dosearch` Algorithm from R for DAGs
 #'
 #' @inheritParams dosearch
+#' @srrstats {NW2.9, NW2.10} Pre-processing routine for DAG inputs.
 #' @noRd
 get_derivation_dag <- function(data, query, graph, transportability,
                                selection_bias, missing_data, control) {
@@ -45,6 +46,7 @@ get_derivation_dag <- function(data, query, graph, transportability,
   args <- parse_input_distributions_dag(args, data, missing_data)
   args <- validate_input_distributions_dag(args)
   args <- validate_query_dag(args)
+  check_graph_size(2L * args$n) # times 2 due to intervention nodes
   check_valid_input(args, control, missing_data)
   res <- initialize_dosearch(
     as.numeric(args$nums[args$dir_lhs]),
@@ -92,6 +94,7 @@ get_derivation_dag <- function(data, query, graph, transportability,
 #' @inheritParams dosearch
 #' @param args A `list` of arguments for `initialize_dosearch`
 #' @param graph The graph as a `character` string.
+#' @srrstats {NW2.6} Validates the input DAG.
 #' @noRd
 transform_graph_dag <- function(args, graph, missing_data) {
   if (!nzchar(graph) && is.null(missing_data)) {
@@ -404,6 +407,7 @@ parse_input_distributions_dag <- function(args, data, missing_data) {
 #'
 #' @param args A `list` of arguments for `initialize_dosearch`.
 #' @param d An `integer` vector of length 4 denoting the distribution.
+#' @srrstats {NW2.6} Validates distributions for DAGs.
 #' @noRd
 validate_distribution_dag <- function(args, msg, d, d_str) {
   left_proxy <- bitwAnd(
@@ -487,6 +491,7 @@ validate_distribution_dag <- function(args, msg, d, d_str) {
 #' Check the Validity of Input Distributions
 #'
 #' @param args A `list` of arguments for `initialize_dosearch`.
+#' @srrstats {NW2.6} Validates input distributions for DAGs.
 #' @noRd
 validate_input_distributions_dag <- function(args) {
   args$p_list <- vector(mode = "list", length = length(args$p_process))
@@ -511,6 +516,7 @@ validate_input_distributions_dag <- function(args) {
 #' Check the Validity of a Target Distribution
 #'
 #' @param args A `list` of arguments for `initialize_dosearch`.
+#' @srrstats {NW2.6} Validates the target distribution for DAGs.
 #' @noRd
 validate_query_dag <- function(args) {
   q <- args$q_process
@@ -533,6 +539,7 @@ validate_query_dag <- function(args) {
 #'
 #' @param args A `list` of arguments for `initialize_dosearch`.
 #' @inheritParams dosearch
+#' @srrstats {NW2.6} Additional checks for potentially mistaken input.
 #' @noRd
 check_valid_input <- function(args, control, missing_data) {
   if (!control$warn) {